+++ CHANGELOG.md
@@ -14,6 +14,8 @@ All notable changes to this project will be documented in this file.
 
 ## [1.0.1.17] - 2019-12-20
 
+### Added
+* function Get-DSM7UserList
 
 ### Fixed